+++ tests/Argument/ScenarioStateArgumentOrganiserTest.php
@@ -9,12 +9,13 @@
  * file that was distributed with this source code.
  */
 
-namespace Gorghoa\ScenarioStateBehatExtension;
+namespace Gorghoa\ScenarioStateBehatExtension\Argument;
 
 use Behat\Testwork\Argument\ArgumentOrganiser;
 use Doctrine\Common\Annotations\Reader;
 use Gorghoa\ScenarioStateBehatExtension\Annotation\ScenarioStateArgument;
 use Gorghoa\ScenarioStateBehatExtension\Context\Initializer\ScenarioStateInitializer;
+use Gorghoa\ScenarioStateBehatExtension\ScenarioStateInterface;
 use Prophecy\Prophecy\ObjectProphecy;
 
 /**
